+++ frontend/src/pages/UserMain/UserMain.tsx
@@ -21,149 +21,6 @@ export interface UserMainState {
 }
 
 const UserMain = (): JSX.Element => {
-  const spaceList: Space[] = [
-    {
-      spaceId: 1,
-      spaceName: '백엔드 강의실',
-      textPosition: 'bottom',
-      color: '#FED7D9',
-      coordinate: {
-        x: 100,
-        y: 90,
-      },
-    },
-    {
-      spaceId: 2,
-      spaceName: '프론트 강의실1',
-      textPosition: 'bottom',
-      color: '#FED7D9',
-      coordinate: {
-        x: 560,
-        y: 40,
-      },
-    },
-    {
-      spaceId: 3,
-      spaceName: '프론트 강의실2',
-      textPosition: 'bottom',
-      color: '#FED7D9',
-      coordinate: {
-        x: 560,
-        y: 140,
-      },
-    },
-    {
-      spaceId: 4,
-      spaceName: '회의실1',
-      textPosition: 'bottom',
-      color: '#FFE3AC',
-      coordinate: {
-        x: 29,
-        y: 229,
-      },
-    },
-    {
-      spaceId: 5,
-      spaceName: '회의실2',
-      textPosition: 'bottom',
-      color: '#FFE3AC',
-      coordinate: {
-        x: 88,
-        y: 229,
-      },
-    },
-    {
-      spaceId: 6,
-      spaceName: '회의실3',
-      textPosition: 'bottom',
-      color: '#FFE3AC',
-      coordinate: {
-        x: 510,
-        y: 220,
-      },
-    },
-    {
-      spaceId: 7,
-      spaceName: '회의실4',
-      textPosition: 'bottom',
-      color: '#FFE3AC',
-      coordinate: {
-        x: 584,
-        y: 220,
-      },
-    },
-    {
-      spaceId: 8,
-      spaceName: '회의실5',
-      textPosition: 'bottom',
-      color: '#FFE3AC',
-      coordinate: {
-        x: 668,
-        y: 335,
-      },
-    },
-    {
-      spaceId: 9,
-      spaceName: '트랙방',
-      textPosition: 'bottom',
-      color: '#D8FBCC',
-      coordinate: {
-        x: 259,
-        y: 336,
-      },
-    },
-    {
-      spaceId: 10,
-      spaceName: '페어룸1',
-      textPosition: 'left',
-      color: '#CCDFFB',
-      coordinate: {
-        x: 208,
-        y: 289,
-      },
-    },
-    {
-      spaceId: 11,
-      spaceName: '페어룸2',
-      textPosition: 'left',
-      color: '#CCDFFB',
-      coordinate: {
-        x: 208,
-        y: 318,
-      },
-    },
-    {
-      spaceId: 12,
-      spaceName: '페어룸3',
-      textPosition: 'left',
-      color: '#CCDFFB',
-      coordinate: {
-        x: 208,
-        y: 347,
-      },
-    },
-    {
-      spaceId: 13,
-      spaceName: '페어룸4',
-      textPosition: 'left',
-      color: '#CCDFFB',
-      coordinate: {
-        x: 208,
-        y: 376,
-      },
-    },
-    {
-      spaceId: 14,
-      spaceName: '페어룸5',
-      textPosition: 'left',
-      color: '#CCDFFB',
-      coordinate: {
-        x: 208,
-        y: 404,
-      },
-    },
-  ];
-
   // Note: 루터회관 14층으로 상정하고 구현. 추후 useSpaces로 대체 필요
   const mapId = 1;
